fix/bid: BidService와 PointService 간의 순환 참조 해결 (Resolve circular dependency between BidService and PointService)
