+++ src/main/java/org/example/lastcall/domain/bid/controller/BidController.java
@@ -6,7 +6,8 @@ import org.example.lastcall.common.security.Auth;
 import org.example.lastcall.domain.auth.model.AuthUser;
 import org.example.lastcall.domain.bid.dto.response.BidGetAllResponse;
 import org.example.lastcall.domain.bid.dto.response.BidResponse;
-import org.example.lastcall.domain.bid.service.BidService;
+import org.example.lastcall.domain.bid.service.command.BidCommandService;
+import org.example.lastcall.domain.bid.service.query.BidQueryService;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Sort;
 import org.springframework.data.web.PageableDefault;
@@ -24,12 +25,13 @@ import lombok.RequiredArgsConstructor;
 @RequiredArgsConstructor
 @RequestMapping("/api/v1/auctions/{auctionId}/bids")
 public class BidController {
-	private final BidService bidService;
+	private final BidQueryService bidQueryService;
+	private final BidCommandService bidCommandService;
 
 	// 입찰 등록
 	@PostMapping
 	public ResponseEntity<ApiResponse<BidResponse>> createBid(@PathVariable Long auctionId, @Auth AuthUser authUser) {
-		BidResponse bid = bidService.createBid(auctionId, authUser);
+		BidResponse bid = bidCommandService.createBid(auctionId, authUser);
 
 		return ResponseEntity.status(HttpStatus.CREATED).body(ApiResponse.success("입찰이 완료되었습니다.", bid));
 	}
@@ -39,7 +41,7 @@ public class BidController {
 	public ResponseEntity<ApiResponse<PageResponse<BidGetAllResponse>>> getAllBids(
 		@PathVariable Long auctionId,
 		@PageableDefault(size = 5, sort = "createdAt", direction = Sort.Direction.DESC) Pageable pageable) {
-		PageResponse<BidGetAllResponse> bids = bidService.getAllBids(auctionId, pageable);
+		PageResponse<BidGetAllResponse> bids = bidQueryService.getAllBids(auctionId, pageable);
 
 		return ResponseEntity.ok(ApiResponse.success("해당 경매의 입찰 내역을 조회합니다.", bids));
 	}
